Extract executor to separate file
